Refactorizar las plantillas de informes de ventas y eliminar la especificación de zona horaria del formato de fecha.
