+++ src/reports/templates/sales/cabeceras/sales-report-header.template.ts
@@ -645,8 +645,6 @@ function formatOptionalDate(value: string | Date | null | undefined): string {
  */
 function formatDate(date: Date): string {
   return new Intl.DateTimeFormat("es-BO", {
-    timeZone: "America/La_Paz",
-
     day: "2-digit",
 
     month: "2-digit",
@@ -660,8 +658,6 @@ function formatDate(date: Date): string {
  */
 function formatTime(date: Date): string {
   return new Intl.DateTimeFormat("es-BO", {
-    timeZone: "America/La_Paz",
-
     hour: "2-digit",
 
     minute: "2-digit",
