+++ src/reports/templates/sales/reports/reportSales.ts
@@ -347,8 +347,6 @@ function formatDateTime(value: string | Date | null | undefined): string {
   }
 
   return new Intl.DateTimeFormat("es-BO", {
-    timeZone: "America/La_Paz",
-
     day: "2-digit",
 
     month: "2-digit",
